refactor(architecture): split editor core and extend plugin seams
Decompose OutlineEditor into outline-editor modules with a shared visible-tree walk, move todo fade/backspace/slash behavior into plugin seams (rowDecorations, caretKeys, core-slash MenuSpec), and colocate tag-colors with the tags plugin.
